requirements/local.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-model-utils==4.3.1
Pillow==9.4.0
pkg==0.2
psycopg2==2.8.4
psycopg2-binary==2.9.5
pytz==2019.3
sqlparse==0.3.1
tzdata==2022.7
Unipath==1.1